code error related with "extern* App" solved
extern means that a variable declared in another place can be used here. if I try to use this extern variable from another file that is not the original file where the variable is declared and defined or the file where is declared with the extern, I won't be able to access intern app class functions and variables. This has been solved by including app.h in every module.cpp. In the headers cant be included because we cause a error that says app has been already d3eclarated and we are redefining it